requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlmodel>=0.0.14
numpy>=1.26.0
pydantic>=2.5.0
litellm>=1.0.0
requests>=2.31.0
fpdf2>=2.7.0
# Note: garak and pyrit can be heavy, for this hackathon we will stub their inputs first to ensure the pipeline works seamlessly
